backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
drf-spectacular==0.27.2
django-filter==23.5
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
python-decouple==3.8
django-storages==1.14.2
whitenoise==6.6.0
django-ratelimit==4.1.0
sendgrid==6.11.0
django-extensions==3.2.3
coverage==7.4.1
pytest-django==4.7.0
factory-boy==3.3.0